Create 1
# pentastagiu_modul1_test
## First test

1. What is regression? What is retesting?
Re-Testing: After a defect is detected and fixed, the software should be retested to confirm that the original defect has been successfully removed. This is done to make sure that the tests cases which failed in last execution are passing after the defects against those failures are fixed.

Regression testing is a type of software testing that tests all area of code or functionalities which can be affected during the solving of a bug and ensure that the new code has not affected other parts of the software.



2. What is the relationship between the following: error, bug, defect, fault, failure?
These terms are an integral part of the software testing process and without detecting and identifying them the team of testers cannot validate the quality, effectiveness, functionality, and more of the software.
During the process of software testing, errors are the most basic discrepancies found by the team of testers.

Name the difference between agile & v-model. (Please specify some roles and cycles for these)

Please state what is black box testing? How about white box testing? You can also give an example

In the images folder you have a login form. Create Test Cases for the given form.

Please state the testing types(functional, non-functional, black box, white box, design technique) for each TC created.

Consider that you have found a defect in the form for the previous exercise. Create a bug report for it